extract movepicker & history to top-level modules
