+++ src/search.rs
@@ -19,17 +19,14 @@ use crate::{
                 is_game_theoretic_score, mate_in, mated_in, tb_loss_in, tb_win_in, MATE_SCORE,
                 MINIMUM_TB_WIN_SCORE,
             },
-            movegen::{
-                self,
-                movepicker::{MovePicker, Stage, WINNING_CAPTURE_SCORE},
-                MoveListEntry, MAX_POSITION_MOVES,
-            },
+            movegen::{self, MoveListEntry, MAX_POSITION_MOVES},
             Board,
         },
         chessmove::Move,
         piece::{Colour, PieceType},
         CHESS960,
     },
+    movepicker::{MovePicker, Stage, WINNING_CAPTURE_SCORE},
     search::pv::PVariation,
     searchinfo::SearchInfo,
     tablebases::{self, probe::WDL},
